+++ SOBAT/package.json
@@ -9,9 +9,6 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@react-native-firebase/app": "^21.5.0",
-    "@react-native-firebase/auth": "^21.5.0",
-    "@react-native-picker/picker": "^2.9.0",
     "@react-navigation/native": "^7.0.2",
     "@react-navigation/native-stack": "^7.0.2",
     "expo": "~52.0.7",
